Fix ComposeAndReduce/PairProduct allocation regression; add diagnostics and Karatsuba optimizations
ComposeAndReduce: replace per-call vector allocation with static flat
array using dirty-index cleanup pattern, fixing 2.9x regression on large
moduli (spec 2897: 31s→10s). PairProduct: same pattern with per-level
static arrays to handle recursion.

Increase canonical set initial buckets from 1M to 4M to reduce early
rehashing. Add 16MB stack via linker flag. Add printProcessMemoryUsage()
(Windows API) and improved catch blocks with typeid diagnostics in
return_map_T.h. Reformulate Karatsuba to use 2 additions instead of 4;
kill dead CFLOBDDs early; guard debug printing with TRACK_PROCESS_MEMORY_USAGE.

+++ CFLOBDD/cross_product.h
@@ -90,7 +90,7 @@ class PairProductMapBody {//: public List<intpair> {
   bool operator==(const PairProductMapBody &p) const;
   intpair& operator[](unsigned int i);                       // Overloaded []
   unsigned int Size();
-  size_t hashCheck;
+  unsigned int hashCheck;
  public:
   bool isCanonical;              // Is this PairProductMapBody in *canonicalPairProductMapBodySet?
   static Hashset<PairProductMapBody> *canonicalPairProductMapBodySet;
